Update README.md
First of all, I introduce myself as Hoztes, I just want to clarify that I am not a programmer, likewise, I do not know much about the theory of social choice, I was just browsing YouTube and I saw a video about this branch of mathematics and I wondered how a fair voting system could be where transparency is implemented, Duverger's law is avoided, the Condorcet loop is avoided, direct confrontations are used and, in case of a tie, a system is implemented to choose the winner (in this case I implemented a system of net victories as a tiebreaker criterion), in addition to being based on Arrow's theorem (I read all this and it caught my attention), after getting the idea I consulted with AI's to solve this problem, I did it more than anything so that in case I have doubts with my friends about what decisions to make we have a tool that is coherent and uses logic, I do not know if this voting system is perfect, I know that It faces drawbacks such as complexity for the voter, and it's not a strategy-testing tool. However, I believe the community can help me improve this. Likewise, I hope you'll lend me your support to develop the idea and be able to implement it in the future, whether for important decisions or simply to choose what to eat while I'm with my friends.